requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-dateutil
imapclient
PyPDF2
bandit
python-multipart
email-validator
python-jose
